ui/requirements-ui.txt
streamlit==1.38.0
ultralytics==8.3.20
pillow==10.4.0
pandas==2.2.2
opencv-python==4.10.0.84
mlflow